move delete method to grid test case and remove redundant methods while form is not ready
