+++ tests/Integration/PrestaShopBundle/Controller/GridControllerTestCase.php
@@ -239,6 +239,38 @@ abstract class GridControllerTestCase extends WebTestCase
     }
 
     /**
+     * @param string $deleteRoute
+     * @param array $routeParams
+     */
+    protected function deleteEntityFromPage(string $deleteRoute, array $routeParams): void
+    {
+        // performs the deletion and then redirects to the list
+        $this->client->request('POST', $this->router->generate($deleteRoute, $routeParams));
+        $this->assertResponseRedirects();
+    }
+
+    /**
+     * @param string $route
+     * @param array<string, mixed> $routeParams
+     */
+    protected function toggleStatus(string $route, array $routeParams): void
+    {
+        $this->client->request('POST', $this->router->generate($route, $routeParams));
+        $this->assertResponseRedirects();
+    }
+
+    /**
+     * @param string $route
+     * @param array $requestParams
+     */
+    protected function bulkDeleteEntitiesFromPage(string $route, array $requestParams): void
+    {
+        // Delete url performs the deletion and then redirects to the list
+        $this->client->request('POST', $this->router->generate($route), $requestParams);
+        $this->assertResponseRedirects();
+    }
+
+    /**
      * Returns the selector allowing to get the grid's search button.
      *
      * @return string
